move dpluger command to its own file
